# docs/swagger.yaml
openapi: 3.0.0
servers:
  - url: "https://api.eesast.com"
    description: Remote API Server
  - url: "http://localhost:{port}"
    description: Local API Server
    variables:
      port:
        description: 端口号
        default: "28888"
info:
  title: API Backend for EESAST
  description: |
    清华大学电子工程系学生科协后端 API。了解更多 EESAST 内容，请访问：[GitHub 主页](https://github.com/eesast/)或 [EESAST 官方网站](https://eesast.com/)。
  version: 1.0.0
  contact:
    name: API Support
    email: i@raineggplant.com
...
paths:
...
  "/v1/teams":
    get:
      tags:
        - teams
      summary: 返回匹配查询的队伍
      description: 可以通过 query 进行筛选，已登录用户可以查询本队邀请码， `root` 角色可查询所有邀请码
      operationId: findTeams
      security:
        - bearerAuth: []
      parameters:
        - name: contestId
          in: query
          description: 比赛ID
          schema:
            type: number
        - name: available
          in: query
          description: 队伍是否仍具有参赛资格
          schema:
            type: boolean
        - name: self
          in: query
          description: 是否只查询含有自身的队伍
          schema:
            type: boolean
        - name: begin
          in: query
          description: 按时间降序排列，起始队伍的位置
          schema:
            type: number
        - name: end
          in: query
          description: 按时间降序排列，结束队伍的位置
          schema:
            type: number
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
        400:
          $ref: "#/components/responses/BadRequest"
        500:
          description: Server error

    post:
      tags:
        - teams
      summary: 新建队伍
      description: 需要已登录用户
      operationId: addTeam
      security:
        - bearerAuth: []
      requestBody:
        $ref: "#/components/requestBodies/addTeam"
      responses:
        201:
          description: Successful operation
          headers:
            Location:
              description: 新建队伍的 URL
              schema:
                type: string
        400:
          # TODO: Solve status code conflict
          # $ref: "#/components/responses/BadRequest"
          description: "TODO: Solve status code conflict. --- Contest not available / User not in track"
        409:
          description: Team name already exists / User is already in a team
        500:
          description: Server error
  "/v1/teams/{teamId}":
    get:
      tags:
        - teams
      summary: 根据 ID 获取队伍
      description: 本队用户或 `root` 角色可见邀请码
      operationId: getTeamById
      security:
        - bearerAuth: []
      parameters:
        - name: teamId
          in: path
          description: 队伍 ID
          required: true
          schema:
            type: number
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        400:
          $ref: "#/components/responses/BadRequest"
        404:
          description: "Not Found: Team does not exist"
        500:
          description: Server error

    put:
      tags:
        - teams
      summary: 更新已存在的队伍
      description: 需要 `root` 角色，或待更新队伍的队长为请求者
      operationId: updateTeam
      security:
        - bearerAuth: []
      parameters:
        - name: teamId
          in: path
          description: 待更新队伍的 ID
          required: true
          schema:
            type: number
      requestBody:
        $ref: "#/components/requestBodies/updateTeam"
      responses:
        204:
          description: Successful operation
          headers:
            Location:
              description: 被更新队伍的 URL
              schema:
                type: string
        400:
          # TODO: Solve status code conflict
          # $ref: "#/components/responses/BadRequest"
          description: "TODO: Solve status code conflict. --- Invalid members / Captain is not a member of the team / Contest not available"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        404:
          description: "Not Found: Team does not exist"
        409:
          description: Team name already exists
        500:
          description: Server error

    delete:
      tags:
        - teams
      summary: 删除一支队伍
      description: 需要 `root` 角色，或待删除队伍的队长为请求者
      operationId: deleteTeam
      security:
        - bearerAuth: []
      parameters:
        - name: teamId
          in: path
          description: 要删除队伍的 ID
          required: true
          schema:
            type: number
      responses:
        204:
          description: Successful operation
        400:
          # TODO: Solve status code conflict
          # $ref: "#/components/responses/BadRequest"
          description: "TODO: Solve status code conflict. --- Contest not available"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        404:
          description: "Not Found: Team does not exist"
        500:
          description: Server error
  "/v1/teams/{teamId}/members":
    get:
      tags:
        - teams
      summary: 根据队伍 ID 获取队员
      operationId: getMembersByTeamId
      parameters:
        - name: teamId
          in: path
          description: 队伍 ID
          required: true
          schema:
            type: number
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: number
        400:
          $ref: "#/components/responses/BadRequest"
        404:
          description: "Not Found: Team does not exist"
        500:
          description: Server error

    post:
      tags:
        - teams
      summary: 根据队伍 ID 向队伍增加队员
      description: 需要 `root` 角色，或增加队员为请求者自身且邀请码正确
      operationId: addMemberToTeamId
      security:
        - bearerAuth: []
      parameters:
        - name: teamId
          in: path
          description: 队伍 ID
          required: true
          schema:
            type: number
      requestBody:
        $ref: "#/components/requestBodies/addTeamMember"
      responses:
        201:
          description: Successful operation
          headers:
            Location:
              description: 新增队员的 URL
              schema:
                type: string
        400:
          # TODO: Solve status code conflict
          # $ref: "#/components/responses/BadRequest"
          description: "TODO: Solve status code conflict. --- Member does not exist / Contest not available / User not in track"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        403:
          description: Team is not available / Incorrect invite code
        404:
          description: "Not Found: Team does not exist"
        409:
          description: The number of members exceeds / User is already in a team
        422:
          description: Missing credentials
        500:
          description: Server error
  "/v1/teams/{teamId}/members/{memberId}":
    delete:
      tags:
        - teams
      summary: 根据ID删除指定队伍中的队员
      description: 需要 `root` 角色，或队伍队长为请求者自身
      operationId: deleteMemberFromTeamId
      security:
        - bearerAuth: []
      parameters:
        - name: teamId
          in: path
          description: 队伍 ID
          required: true
          schema:
            type: number
        - name: memberId
          in: path
          description: 队员 ID
          required: true
          schema:
            type: number
      responses:
        204:
          description: Successful operation
        400:
          # TODO: Solve status code conflict
          # $ref: "#/components/responses/BadRequest"
          description: "TODO: Solve status code conflict. --- Leader cannot be deleted / Contest not available"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        404:
          description: "Not Found: Team does not exist / Member does not exist"
        500:
          description: Server error
  "/v1/teams/scores":
    put:
      tags:
        - teams
      summary: 根据比赛结果更新多支队伍分数
      description: 由server发起
      operationId: updateTeamsScore
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - scores
              properties:
                scores:
                  type: array
      responses:
        204:
          description: Successful operation
          headers:
            Location:
              description: 被更新队伍的 URL
              schema:
                type: string
        401:
          $ref: "#/components/responses/UnauthorizedError"
        403:
          description: "Forbidden: Permission denied"
        404:
          description: "Not Found: Team does not exist"
        422:
          description: "Unprocessable Entity: Missing form data"
        500:
          description: Server error
  "/v1/teams/{teamId}/score":
    put:
      tags:
        - teams
      summary: 根据id更新队伍分数
      description: 由server发起
      operationId: updateTeamScore
      security:
        - bearerAuth: []
      parameters:
        - name: teamId
          in: path
          description: 队伍ID
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - score
              properties:
                score:
                  type: number
                replace:
                  type: boolean
      responses:
        204:
          description: Successful operation
          headers:
            Location:
              description: 被更新队伍的 URL
              schema:
                type: string
        401:
          $ref: "#/components/responses/UnauthorizedError"
        403:
          description: "Forbidden: Permission denied"
        404:
          description: "Not Found: Team does not exist"
        422:
          description: "Unprocessable Entity: Missing form data"
        500:
          description: Server error
...
components:
  schemas:
...
    Date:
      type: string
      format: date-time
      description: >-
        日期对象。使用时采用 ISO
        格式字符串表示。[更多信息](https://www.iso.org/iso-8601-date-and-time-format.html)
    Team:
      type: object
      required:
        - id
        - contestId
        - name
        - description
        - leader
        - members
        - inviteCode
      properties:
        id:
          type: number
          description: 自增值
        contestId:
          type: number
          description: 比赛ID
        name:
          type: string
        description:
          type: string
          description: 队伍简介
        leader:
          type: number
          description: 队长ID
        members:
          type: array
          items:
            type: number
            description: 队员ID
        inviteCode:
          type: number
        available:
          type: boolean
          default: true
          description: 队伍是否仍具有参赛资格
        score:
          type: number
          description: 队伍分数
        createdAt:
          $ref: "#/components/schemas/Date"
        createdBy:
          type: number
        updatedAt:
          $ref: "#/components/schemas/Date"
        updatedBy:
          type: number
...
  requestBodies:
...
    addTeam:
      content:
        application/json:
          schema:
            required:
              - contestId
              - name
              - description
            properties:
              contestId:
                type: number
              name:
                type: string
              description:
                type: string
              leader:
                type: number
              members:
                type: array
                items:
                  type: number
              available:
                type: boolean
      description: 新建队伍的提交格式
      required: true
    updateTeam:
      content:
        application/json:
          schema:
            properties:
              name:
                type: string
              description:
                type: string
              leader:
                type: number
              members:
                type: array
                items:
                  type: number
              available:
                type: boolean
              score:
                type: number
      description: 更新队伍的提交格式
      required: true
    addTeamMember:
      content:
        application/json:
          schema:
            required:
              - id
              - inviteCode
            properties:
              id:
                type: number
              inviteCode:
                type: string
      description: 增加队员的提交格式
      required: true
...
  responses:
    BadRequest:
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: array
                description: 错误列表
                items:
                  type: object
                  properties:
                    path:
                      type: string
                      description: 错误位置
                      example: ".params.articleId"
                    message:
                      type: string
                      description: 错误信息
                      example: "should be number"
                    errorCode:
                      type: string
                      description: 错误码
                      example: "type.openapi.validation"
      description: 包含请求错误信息的 JSON
    UnauthorizedError:
      description: "Access token is missing, invalid or expired / Permission denied"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT